L2TP: Add preference to select Ericsson HDLC flavor
At some point over the wireshark last years, the connection between the
L2TP dissector and the Ericsson HDLC dissector (packet-ehdlc.c) got
broken.  I cannot poinpoint the exact change when it broke.  However,
traces with Ericsson L2TP that used to display correctly no longer do
so.

The problem with Ericsson L2TP is that they simply re-use the HDLC
pseudowire (0x0006), but then have frames completely different than
what's specified in the RFCs for that.

To address this, let's add a protocol preference to allow the user to
choose.

Change-Id: Idce022ef0f233c01ecffc67dca7f428bde31ef6b