l2tp: Introduce preference for default pseudowire
In some cases, the establishment of the L2TP session is not captured in
the pcap file, and thus wireshark cannot determine the l2tp pseudowire
to be used for decode.

So rather than using a ocmpile-time hard-coded default of cisco hdlc,
let's introduce a prference where the user can select the default
pseudowire in such cases.

Change-Id: I44a6373d4812c9ab5a919514273a41a2ab57142c